+++ CHANGELOG.md
@@ -40,7 +40,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Bug Fixes
 
-* (ante) [tharsis#964](https://github.com/tharsis/ethermint/pull/964) fix gas meter reset with user provided gas limit in the innermost EthAnteHandler. Fixes block's consumed gas calculation in the block creation phase.
+* (ante) [tharsis#964](https://github.com/tharsis/ethermint/pull/964) add NewInfiniteGasMeterWithLimit for storing the user provided gas limit. Fixes block's consumed gas calculation in the block creation phase.
 * (rpc) [#970] (https://github.com/tharsis/ethermint/pull/970) Fix unexpected nil reward values on `eth_feeHistory` response
 * (evm) [\#529](https://github.com/tharsis/ethermint/issues/529) support return value on trace tx response.